refactor(general): move source to src folder
