+++ package.json
@@ -2,7 +2,7 @@
   "name": "express-implhandler",
   "version": "2.0.1",
   "description": "Simple express method implementation script to catch error code 501",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "lint": "eslint --ignore-path .gitignore .",
     "test": "mocha --exit",
